add view model & data binding library
