+++ front-mobile/app/src/main/java/com/project/frontMobile/ui/authentication/forgotpassword/ForgotPasswordFragment.kt
@@ -1,0 +1,25 @@
+package com.project.frontMobile.ui.authentication.forgotpassword
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.project.frontMobile.R
+
+class ForgotPasswordFragment : Fragment() {
+    // TODO: Rename and change types of parameters
+    private var param1: String? = null
+    private var param2: String? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_forgot_password, container, false)
+    }
+}
