+++ front-mobile/app/src/main/java/com/project/frontMobile/ui/author/AuthorFragment.kt
@@ -1,0 +1,22 @@
+package com.project.frontMobile.ui.author
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.project.frontMobile.R
+
+class AuthorFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_author, container, false)
+    }
+}
