+++ front-mobile/app/src/main/java/com/project/frontMobile/ui/rgpd/RGPDFragment.kt
@@ -1,0 +1,22 @@
+package com.project.frontMobile.ui.rgpd
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.project.frontMobile.R
+
+class RGPDFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_r_g_p_d, container, false)
+    }
+}
