--- front-mobile/app/src/main/java/com/project/frontMobile/view/MainActivity.kt
@@ -1,45 +1,0 @@
-package com.project.frontMobile.view
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.NavigationUI
-import androidx.navigation.ui.setupActionBarWithNavController
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.project.frontMobile.R
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var navController: NavController
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        setupNav()
-    }
-
-    fun setupNav() {
-        val bottomNavView = findViewById<BottomNavigationView>(R.id.bottom_navigation_view)
-        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        navController = navHostFragment.navController
-
-        NavigationUI.setupWithNavController(bottomNavView, navController)
-
-        navController.addOnDestinationChangedListener { controller, destination, arguments ->
-            when (destination.id) {
-                R.id.libraryFragment -> bottomNavView.visibility = View.VISIBLE
-                R.id.upcomingFragment -> bottomNavView.visibility = View.VISIBLE
-                R.id.suggestionFragment -> bottomNavView.visibility = View.VISIBLE
-                R.id.menuFragment -> bottomNavView.visibility = View.VISIBLE
-                else -> bottomNavView.visibility = View.GONE
-            }
-        }
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return navController.navigateUp() || super.onSupportNavigateUp()
-    }
-}
