--- front-mobile/app/src/main/java/com/project/frontMobile/view/settings/SettingsFragment.kt
@@ -1,24 +1,0 @@
-package com.project.frontMobile.view.settings
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.findNavController
-import com.google.android.material.button.MaterialButton
-import com.project.frontMobile.R
-
-class SettingsFragment : Fragment() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_settings, container, false)
-    }
-}
