+++ front-mobile/app/src/main/res/layout/fragment_authentication.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".view.authentication.AuthenticationFragment">
+    tools:context=".ui.authentication.AuthenticationFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/log_in_button"
