+++ front-mobile/app/src/main/res/layout/fragment_author.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.author.AuthorFragment">
+    tools:context=".ui.author.AuthorFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
