+++ front-mobile/app/src/main/res/layout/fragment_forgot_password.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.authentication.forgotpassword.ForgotPasswordFragment">
+    tools:context=".ui.authentication.forgotpassword.ForgotPasswordFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
