+++ front-mobile/app/src/main/res/layout/fragment_library.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.library.LibraryFragment">
+    tools:context=".ui.library.LibraryFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/book_button"
