+++ front-mobile/app/src/main/res/layout/fragment_loading.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".view.loading.LoadingFragment">
+    tools:context=".ui.loading.LoadingFragment">
 
     <com.google.android.material.progressindicator.CircularProgressIndicator
         android:layout_width="wrap_content"
