+++ front-mobile/app/src/main/res/layout/fragment_log_in.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.authentication.login.LogInFragment">
+    tools:context=".ui.authentication.login.LogInFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/log_in_button"
