+++ front-mobile/app/src/main/res/layout/fragment_menu.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.menu.MenuFragment">
+    tools:context=".ui.menu.MenuFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/profile_button"
