+++ front-mobile/app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.user.ProfileFragment">
+    tools:context=".ui.user.ProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
