+++ front-mobile/app/src/main/res/layout/fragment_r_g_p_d.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.rgpd.RGPDFragment">
+    tools:context=".ui.rgpd.RGPDFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
