+++ front-mobile/app/src/main/res/layout/fragment_settings.xml
@@ -5,6 +5,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.settings.SettingsFragment">
+    tools:context=".ui.settings.SettingsFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
