+++ front-mobile/app/src/main/res/layout/fragment_sign_up.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".view.authentication.signup.SignUpFragment">
+    tools:context=".ui.authentication.signup.SignUpFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/sign_up_button"
