+++ front-mobile/app/src/main/res/layout/fragment_suggestion.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.suggestion.SuggestionFragment">
+    tools:context=".ui.suggestion.SuggestionFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/search_button"
